requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
CoAPthon3==1.0.2
idna==3.3
paho-mqtt==1.5.1
requests==2.26.0
urllib3==1.26.7
